requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.8
docker==5.0.3
Flask==2.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
zipp==3.6.0
